update train on cpu
