--- .travis.old.yml
@@ -1,60 +1,0 @@
-sudo: required
-dist: xenial
-language: python
-
-python:
-  - "3.4"
-
-install:
-
-  ########################
-  ## Instant DeepDetect ##
-  ########################
-  # dd dependencies
-  - sudo apt-get install -y build-essential libgoogle-glog-dev libgflags-dev libeigen3-dev libopencv-dev libcppnetlib-dev libboost-dev libboost-iostreams-dev libcurlpp-dev libcurl4-openssl-dev protobuf-compiler libopenblas-dev libhdf5-dev libprotobuf-dev libleveldb-dev libsnappy-dev liblmdb-dev libutfcpp-dev
-
-  # download libcurlpp-dev
-  - sudo apt-get remove --yes libcurlpp0
-  - git clone https://github.com/datacratic/curlpp
-  - cd curlpp
-  - ./autogen.sh
-  - ./configure --prefix=/usr --enable-ewarning=no
-  - make
-  - sudo make install
-
-  # compile dd
-  - git clone https://github.com/beniz/deepdetect.git $HOME/deepdetect
-  - sudo cp $TRAVIS_BUILD_DIR/FindEigen3.cmake $HOME/deepdetect/cmake/.
-  - cd $HOME/deepdetect && mkdir build && cd build && cmake .. && make && cd main
-  - nohup ./dede --port 8085 >> dede.log &
-  - cd $TRAVIS_BUILD_DIR
-
-
-  #################################
-  ## Install python dependencies ##
-  #################################
-  - if [[ "$TRAVIS_PYTHON_VERSION" == "2.7" ]]; then
-      wget https://repo.continuum.io/miniconda/Miniconda-latest-Linux-x86_64.sh -O miniconda.sh;
-    else
-      wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh;
-    fi
-  - bash miniconda.sh -b -p $HOME/miniconda
-  - export PATH="$HOME/miniconda/bin:$PATH"
-  - hash -r
-  - conda config --set always_yes yes --set changeps1 no
-  - conda update -q conda
-  # Useful for debugging any issues with conda
-  - conda info -a
-
-  - conda create -q -n test-environment python=$TRAVIS_PYTHON_VERSION numpy scipy pandas pytest h5py
-  - source activate test-environment
-  - pip install pytest pytest-cov python-coveralls pytest-xdist coverage pep8 pytest-pep8
-  - pip install -r requirements.txt
-  - python setup.py install
-
-script:
-#  PYTHONPATH=$PWD:$PYTHONPATH py.test --pep8 -m pep8 -n0;
-  PYTHONPATH=$PWD:$PYTHONPATH py.test tests/
-
-after_success:
-  - coveralls
